fix load_from_file error

+++ src/Model/s21_graph_algorithms.cc
@@ -1,7 +1,7 @@
 #include "s21_graph_algorithms.h"
 using namespace s21;
 
-std::vector<size_t> GraphAlgorithms::DepthFirstSearch(Graph& graph, int start_vertex) {
+std::vector<size_t> GraphAlgorithms::DepthFirstSearch(Graph& graph, size_t start_vertex) {
 	if (start_vertex < 1) throw std::invalid_argument("Index of vertex is incorrect");
 	if (start_vertex > graph.GetSize()) throw std::out_of_range("Index of vertex is out of range");
 	std::vector<bool> visited(graph.GetSize());
@@ -28,7 +28,7 @@ std::vector<size_t> GraphAlgorithms::DepthFirstSearch(Graph& graph, int start_ve
 	return result;
 }
 
-std::vector<size_t> GraphAlgorithms::BreadthFirstSearch(Graph& graph, int start_vertex) {
+std::vector<size_t> GraphAlgorithms::BreadthFirstSearch(Graph& graph, size_t start_vertex) {
 	if (start_vertex < 1) throw std::invalid_argument("Index of vertex is incorrect");
 	if (start_vertex > graph.GetSize()) throw std::out_of_range("Index of vertex is out of range");
 	std::vector<bool> visited(graph.GetSize());
@@ -52,7 +52,7 @@ std::vector<size_t> GraphAlgorithms::BreadthFirstSearch(Graph& graph, int start_
 	return result;
 }
 
-size_t s21::GraphAlgorithms::GetShortestPathBetweenVertices(Graph& graph, int vertex1, int vertex2) {
+size_t s21::GraphAlgorithms::GetShortestPathBetweenVertices(Graph& graph, size_t vertex1, size_t vertex2) {
 	if (vertex1 < 1 || vertex2 < 1) throw std::invalid_argument("Index of vertex is incorrect");
 	if (vertex1 > graph.GetSize() || vertex2 > graph.GetSize()) throw std::out_of_range("Index of vertex is out of range");
 	std::vector<size_t> min_length(graph.GetSize(), std::numeric_limits<size_t>::max());
